Move colour definitions to parent component so it can be used in Footer Divider component as well
